Start the Spanish translation of this book
Start the Spanish translation of this book. It is a very good
idea to check out the original book at:
https://github.com/esl/erlang-handbook

This repository has two goals:

1. Learn Erlang
2. Translate this book to Spanish

As you can imagine, I'd like to translate to Spanish this book
at the same time that I learn Erlang.

Contact info: joaquin@isla.io